+++ Src/libCZI_UnitTests/MockMetadataSegment.cpp
@@ -3,6 +3,8 @@
 // SPDX-License-Identifier: LGPL-3.0-or-later
 
 #include "include_gtest.h"
+#include <cstddef>
+#include <sys/_types/_ssize_t.h>
 #include "MockMetadataSegment.h"
 
 using namespace libCZI;
@@ -6778,7 +6780,7 @@ static string GetTestData()
     // the size of static strings seems to be limited (with various compilers), so the easiest workaround was to
     // have a lot of short strings and put them together here...
     string str;
-    for (int i = 0; i < sizeof(testMetadata) / sizeof(testMetadata[0]); ++i)
+    for (size_t i = 0; i < sizeof(testMetadata) / sizeof(testMetadata[0]); ++i)
     {
         str += testMetadata[i];
     }
@@ -6791,7 +6793,7 @@ static string GetTest2Data()
     // the size of static strings seems to be limited (with various compilers), so the easiest workaround was to
     // have a lot of short strings and put them together here...
     string str;
-    for (int i = 0; i < sizeof(test2Metadata) / sizeof(test2Metadata[0]); ++i)
+    for (size_t i = 0; i < sizeof(test2Metadata) / sizeof(test2Metadata[0]); ++i)
     {
         str += test2Metadata[i];
     }
@@ -6804,7 +6806,7 @@ static string GetTest3Data()
     // the size of static strings seems to be limited (with various compilers), so the easiest workaround was to
     // have a lot of short strings and put them together here...
     string str;
-    for (int i = 0; i < sizeof(test3Metadata) / sizeof(test3Metadata[0]); ++i)
+    for (size_t i = 0; i < sizeof(test3Metadata) / sizeof(test3Metadata[0]); ++i)
     {
         str += test3Metadata[i];
     }
@@ -6817,7 +6819,7 @@ static string GetTest4Data()
     // the size of static strings seems to be limited (with various compilers), so the easiest workaround was to
     // have a lot of short strings and put them together here...
     string str;
-    for (int i = 0; i < sizeof(test4Metadata) / sizeof(test4Metadata[0]); ++i)
+    for (size_t i = 0; i < sizeof(test4Metadata) / sizeof(test4Metadata[0]); ++i)
     {
         str += test4Metadata[i];
     }
@@ -6830,7 +6832,7 @@ static string GetTest5Data()
     // the size of static strings seems to be limited (with various compilers), so the easiest workaround was to
     // have a lot of short strings and put them together here...
     string str;
-    for (int i = 0; i < sizeof(test5Metadata) / sizeof(test5Metadata[0]); ++i)
+    for (size_t i = 0; i < sizeof(test5Metadata) / sizeof(test5Metadata[0]); ++i)
     {
         str += test5Metadata[i];
     }
@@ -6843,7 +6845,7 @@ static string GetTest6Data()
     // the size of static strings seems to be limited (with various compilers), so the easiest workaround was to
     // have a lot of short strings and put them together here...
     string str;
-    for (int i = 0; i < sizeof(test6Metadata) / sizeof(test6Metadata[0]); ++i)
+    for (size_t i = 0; i < sizeof(test6Metadata) / sizeof(test6Metadata[0]); ++i)
     {
         str += test6Metadata[i];
     }
@@ -6855,7 +6857,7 @@ static string GetTestDataInvalid()
 {
     // here we leave out the last line, resulting in invalid XML
     string str;
-    for (int i = 0; i < sizeof(testMetadata) / sizeof(testMetadata[0]) - 1; ++i)
+    for (size_t i = 0; i < sizeof(testMetadata) / sizeof(testMetadata[0]) - 1; ++i)
     {
         str += testMetadata[i];
     }
@@ -6891,7 +6893,7 @@ MockMetadataSegment::MockMetadataSegment(Type type/*= Type::Data1*/)
     }
 }
 
-/*virtual*/ std::shared_ptr<const void> MockMetadataSegment::GetRawData(MemBlkType type, size_t* ptrSize)
+/*virtual*/ std::shared_ptr<const void> MockMetadataSegment::GetRawData(MemBlkType, size_t*)
 {
     throw  std::logic_error("Function not implemented.");
 }
